# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FaceTrack API
  description: 'FaceTrack attendance system API (requirements §12). Portal: session
    cookie + CSRF header; integrations: X-API-Key.'
  version: 1.0.0
paths:
...
  /api/v1/locations:
    get:
      tags:
      - organization
      summary: List locations
      operationId: list_locations
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
            pattern: ^-?[a-z_]+$
          - type: 'null'
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_LocationOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - organization
      summary: Create Location
      operationId: create_location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_LocationOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/locations/{location_id}:
    put:
      tags:
      - organization
      summary: Update Location
      operationId: update_location
      parameters:
      - name: location_id
        in: path
        required: true
        schema:
          type: integer
          title: Location Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_LocationOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - organization
      summary: Delete Location
      operationId: delete_location
      parameters:
      - name: location_id
        in: path
        required: true
        schema:
          type: integer
          title: Location Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/departments:
    get:
      tags:
      - organization
      summary: List departments
      operationId: list_departments
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
            pattern: ^-?[a-z_]+$
          - type: 'null'
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_DepartmentOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - organization
      summary: Create Department
      operationId: create_department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DepartmentOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/departments/{department_id}:
    get:
      tags:
      - organization
      summary: Get Department
      operationId: get_department
      parameters:
      - name: department_id
        in: path
        required: true
        schema:
          type: integer
          title: Department Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DepartmentOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - organization
      summary: Update Department
      operationId: update_department
      parameters:
      - name: department_id
        in: path
        required: true
        schema:
          type: integer
          title: Department Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DepartmentOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - organization
      summary: Delete Department
      operationId: delete_department
      parameters:
      - name: department_id
        in: path
        required: true
        schema:
          type: integer
          title: Department Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/shifts:
    get:
      tags:
      - organization
      summary: List shifts
      operationId: list_shifts
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
            pattern: ^-?[a-z_]+$
          - type: 'null'
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_ShiftOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - organization
      summary: Create Shift
      operationId: create_shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ShiftOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/shifts/{shift_id}:
    get:
      tags:
      - organization
      summary: Get Shift
      operationId: get_shift
      parameters:
      - name: shift_id
        in: path
        required: true
        schema:
          type: integer
          title: Shift Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ShiftOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - organization
      summary: Update Shift
      operationId: update_shift
      parameters:
      - name: shift_id
        in: path
        required: true
        schema:
          type: integer
          title: Shift Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ShiftOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - organization
      summary: Delete Shift
      operationId: delete_shift
      parameters:
      - name: shift_id
        in: path
        required: true
        schema:
          type: integer
          title: Shift Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/holidays:
    get:
      tags:
      - organization
      summary: List holidays
      operationId: list_holidays
      parameters:
      - name: location_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Location Id
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date From
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date To
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
            pattern: ^-?[a-z_]+$
          - type: 'null'
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_HolidayOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - organization
      summary: Create Holiday
      operationId: create_holiday
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HolidayCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_HolidayOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/holidays/{holiday_id}:
    put:
      tags:
      - organization
      summary: Update Holiday
      operationId: update_holiday
      parameters:
      - name: holiday_id
        in: path
        required: true
        schema:
          type: integer
          title: Holiday Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HolidayUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_HolidayOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - organization
      summary: Delete Holiday
      operationId: delete_holiday
      parameters:
      - name: holiday_id
        in: path
        required: true
        schema:
          type: integer
          title: Holiday Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiResponse_DepartmentOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/DepartmentOut'
      type: object
      required:
      - message
      - data
      title: ApiResponse[DepartmentOut]
...
    ApiResponse_HolidayOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/HolidayOut'
      type: object
      required:
      - message
      - data
      title: ApiResponse[HolidayOut]
...
    ApiResponse_LocationOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/LocationOut'
      type: object
      required:
      - message
      - data
      title: ApiResponse[LocationOut]
...
    ApiResponse_ShiftOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/ShiftOut'
      type: object
      required:
      - message
      - data
      title: ApiResponse[ShiftOut]
...
    DepartmentCreate:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        code:
          type: string
          maxLength: 32
          minLength: 1
          pattern: ^[A-Za-z0-9_-]+$
          title: Code
        manager_user_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Manager User Id
      additionalProperties: false
      type: object
      required:
      - name
      - code
      title: DepartmentCreate
    DepartmentOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        code:
          type: string
          title: Code
        manager_user_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Manager User Id
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - id
      - name
      - code
      - manager_user_id
      - created_at
      title: DepartmentOut
    DepartmentUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Name
        code:
          anyOf:
          - type: string
            maxLength: 32
            minLength: 1
            pattern: ^[A-Za-z0-9_-]+$
          - type: 'null'
          title: Code
        manager_user_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Manager User Id
      additionalProperties: false
      type: object
      title: DepartmentUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HolidayCreate:
      properties:
        location_id:
          type: integer
          title: Location Id
        date:
          type: string
          format: date
          title: Date
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
      additionalProperties: false
      type: object
      required:
      - location_id
      - date
      - name
      title: HolidayCreate
    HolidayOut:
      properties:
        id:
          type: integer
          title: Id
        location_id:
          type: integer
          title: Location Id
        date:
          type: string
          format: date
          title: Date
        name:
          type: string
          title: Name
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - id
      - location_id
      - date
      - name
      - created_at
      title: HolidayOut
    HolidayUpdate:
      properties:
        date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date
        name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Name
      additionalProperties: false
      type: object
      title: HolidayUpdate
...
    LocationCreate:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        address:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Address
        timezone:
          type: string
          title: Timezone
          default: Asia/Karachi
      additionalProperties: false
      type: object
      required:
      - name
      title: LocationCreate
    LocationOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        timezone:
          type: string
          title: Timezone
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - id
      - name
      - address
      - timezone
      - created_at
      title: LocationOut
    LocationUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Name
        address:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Address
        timezone:
          anyOf:
          - type: string
          - type: 'null'
          title: Timezone
      additionalProperties: false
      type: object
      title: LocationUpdate
...
    PaginatedResponse_DepartmentOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          items:
            $ref: '#/components/schemas/DepartmentOut'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - message
      - data
      - pagination
      title: PaginatedResponse[DepartmentOut]
...
    PaginatedResponse_HolidayOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          items:
            $ref: '#/components/schemas/HolidayOut'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - message
      - data
      - pagination
      title: PaginatedResponse[HolidayOut]
...
    PaginatedResponse_LocationOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          items:
            $ref: '#/components/schemas/LocationOut'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - message
      - data
      - pagination
      title: PaginatedResponse[LocationOut]
    PaginatedResponse_ShiftOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          items:
            $ref: '#/components/schemas/ShiftOut'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - message
      - data
      - pagination
      title: PaginatedResponse[ShiftOut]
...
    Pagination:
      properties:
        page:
          type: integer
          minimum: 1.0
          title: Page
        page_size:
          type: integer
          minimum: 1.0
          title: Page Size
        total:
          type: integer
          minimum: 0.0
          title: Total
      type: object
      required:
      - page
      - page_size
      - total
      title: Pagination
...
    ShiftCreate:
      properties:
        grace_in_min:
          type: integer
          maximum: 240.0
          minimum: 0.0
          title: Grace In Min
          default: 15
        grace_out_min:
          type: integer
          maximum: 240.0
          minimum: 0.0
          title: Grace Out Min
          default: 15
        half_day_pct:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Half Day Pct
          default: 50
        is_night_shift:
          type: boolean
          title: Is Night Shift
          default: false
        weekly_offs:
          items:
            type: integer
          type: array
          maxItems: 7
          title: Weekly Offs
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        start_time:
          type: string
          format: time
          title: Start Time
        end_time:
          type: string
          format: time
          title: End Time
      additionalProperties: false
      type: object
      required:
      - name
      - start_time
      - end_time
      title: ShiftCreate
    ShiftOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        start_time:
          type: string
          format: time
          title: Start Time
        end_time:
          type: string
          format: time
          title: End Time
        grace_in_min:
          type: integer
          title: Grace In Min
        grace_out_min:
          type: integer
          title: Grace Out Min
        half_day_pct:
          type: integer
          title: Half Day Pct
        is_night_shift:
          type: boolean
          title: Is Night Shift
        weekly_offs:
          items:
            type: integer
          type: array
          title: Weekly Offs
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - id
      - name
      - start_time
      - end_time
      - grace_in_min
      - grace_out_min
      - half_day_pct
      - is_night_shift
      - weekly_offs
      - created_at
      title: ShiftOut
    ShiftUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Name
        start_time:
          anyOf:
          - type: string
            format: time
          - type: 'null'
          title: Start Time
        end_time:
          anyOf:
          - type: string
            format: time
          - type: 'null'
          title: End Time
        grace_in_min:
          anyOf:
          - type: integer
            maximum: 240.0
            minimum: 0.0
          - type: 'null'
          title: Grace In Min
        grace_out_min:
          anyOf:
          - type: integer
            maximum: 240.0
            minimum: 0.0
          - type: 'null'
          title: Grace Out Min
        half_day_pct:
          anyOf:
          - type: integer
            maximum: 100.0
            minimum: 1.0
          - type: 'null'
          title: Half Day Pct
        is_night_shift:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Night Shift
        weekly_offs:
          anyOf:
          - items:
              type: integer
            type: array
          - type: 'null'
          title: Weekly Offs
      additionalProperties: false
      type: object
      title: ShiftUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError